requirements.txt
# Image processing
Pillow>=10.0.0

# SVG to PNG conversion (requires system cairo library)
# Install system deps first: ./install_cairo.sh
cairosvg>=2.7.0

# Clipboard support
pyperclip>=1.8.2

# HTTP requests
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0
